Updated to include weight. Also only using the last months worth of data unless FULL_GRAB is set

+++ beats.py
@@ -13,12 +13,14 @@ import oauth
 logging.basicConfig(level=logging.DEBUG, format="%(asctime)s {%(pathname)s:%(lineno)d} %(levelname)s - %(message)s")
 
 OATH_TOKEN_PATH = os.environ.get("OATH_FILE_LOCATION", "fitbit_oauth.json")
-CACHE_DIR = "./.cache/"
+CACHE_DIR = os.environ.get("BEATS_CACHE_LOCATION", "./.cache/")
 
 # Get the oldest date. Options are hard coded here, or env['DATE_OF_PURCHSE'] in iso format.
 OLDEST_DATE = os.environ.get("DATE_OF_PURCHSE", "2022-01-01T00:00:00")
 OLDEST_DATE = datetime.datetime.fromisoformat(OLDEST_DATE)
 
+FULL_GRAB =  os.environ.get("BEATS_FULL_DATE", "False")
+FULL_GRAB = FULL_GRAB.lower() == "true"
 
 def dump_token(data, token):
     logging.info(f"Writing token {token} to file {OATH_TOKEN_PATH}")
@@ -113,8 +115,12 @@ def collect_activity(fitbit_client: fitbit.Fitbit, activity: str) -> Dict:
     daily = []
     if not os.path.exists(cache_file):
         key = activity.replace("/", "-")
-        for from_date, to_date in date_generator_99_days():
-            daily += fitbit_client.time_series(activity, base_date=from_date, end_date=to_date)[key]
+        if FULL_GRAB:
+            for from_date, to_date in date_generator():
+                daily += fitbit_client.time_series(activity, base_date=from_date, end_date=to_date)[key]
+        else:
+            from_date, to_date = date_month()
+            daily = fitbit_client.time_series(activity, base_date=from_date, end_date=to_date)[key]
         open(cache_file, "w").write(json.dumps(daily))
     else:
         logging.info("Using cache!")
@@ -122,13 +128,19 @@ def collect_activity(fitbit_client: fitbit.Fitbit, activity: str) -> Dict:
     return daily
 
 
-def date_generator_99_days():
-    """generate date ranges in chunks of 99 days from OLDEST_DATE to today
+def date_month():
+    to_date = datetime.datetime.today()
+    from_date = to_date - datetime.timedelta(days=31)
+    return from_date, to_date
 
+
+def date_generator():
+    """
+    generate date ranges in chunks of 99 days from OLDEST_DATE to today
     :yield: (date1, date2)
     """
-    from_date = OLDEST_DATE
     to_date = datetime.datetime.today()
+    from_date = OLDEST_DATE
     while from_date <= to_date:
         yield from_date, from_date + datetime.timedelta(days=99)
         from_date = from_date + datetime.timedelta(days=99)
@@ -183,17 +195,29 @@ def main():
         verify_ssl=False,
     )
 
-    # Drop and recreate the DB. TODO incrementally update?
-    db.drop_database(dbname)
-    db.create_database(dbname)
-    # One shard per year.
-    db.alter_retention_policy("autogen", dbname, "INF", 1, shard_duration="52w")
+    if FULL_GRAB:
+        # Drop and recreate the DB. TODO incrementally update?
+        db.drop_database(dbname)
+        db.create_database(dbname)
+        # One shard per year.
+        db.alter_retention_policy("autogen", dbname, "INF", 1, shard_duration="52w")
+
+    # Weight first
+    raw_data = collect_activity(fitbit_client, "body/weight")
+    if len(raw_data) > 0:
+        first_day = datetime.datetime.fromisoformat(raw_data[0]["dateTime"])
+        daterange = pd.date_range(start=first_day, periods=len(raw_data), freq="D")
+        data = [float(x["value"])  * 0.453592 for x in raw_data] # Convert to kgs
+        df = pd.DataFrame(data=data, index=daterange)
+        write_to_influxdb(db, df, "weight", dbname)
+    else:
+        logging.info("No weight data this month")
 
     # Heart rate
     raw_data = collect_activity(fitbit_client, "activities/heart")
     first_day = datetime.datetime.fromisoformat(raw_data[0]["dateTime"])
     daterange = pd.date_range(start=first_day, periods=len(raw_data), freq="D")
-    data = [x["value"].get("restingHeartRate", None) for x in raw_data]
+    data = [float(x["value"].get("restingHeartRate", None)) for x in raw_data]
     df = pd.DataFrame(data=data, index=daterange)
     write_to_influxdb(db, df, "resting_hr", dbname)
 
@@ -212,12 +236,20 @@ def main():
     raw_data = nap_filter(raw_data_unfiltered)
 
     # Date range should be from day of purchase to today
-    first_day = datetime.datetime.fromisoformat(raw_data[-1]["dateOfSleep"])  # REVERSE ORDER WTF FITBIT
+    # Get the earliest date of sleep, using a list comprehension
+    first_day = min([datetime.datetime.fromisoformat(x["dateOfSleep"]) for x in raw_data])
+
     daterange = pd.date_range(start=first_day, periods=len(raw_data), freq="D")
 
     # Start writing data
     # Write ones that dont require manipulating types
-    for sleep_type in ["awakeCount", "awakeningsCount", "efficiency", "minutesAsleep"]:
+    for sleep_type in ["efficiency"]:
+        data = [float(x[sleep_type]) for x in raw_data]
+        df = pd.DataFrame(data=data, index=daterange)
+        write_to_influxdb(db, df, f"sleep_{sleep_type}", dbname)
+        
+    # Same as above, but not floats. Not sure why 
+    for sleep_type in ["awakeCount","awakeningsCount", "minutesAsleep"]:
         data = [x[sleep_type] for x in raw_data]
         df = pd.DataFrame(data=data, index=daterange)
         write_to_influxdb(db, df, f"sleep_{sleep_type}", dbname)
